+++ CHANGES.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - use sharing when building canonical types and terms
 - add option --use-sharing for using sharing in lp output when defining term abbreviations
 - command nbp to get the number of proof steps
+- commands patch, unpatch and link to call the corresponding scripts
 
 ### Modified